backend/requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
bleach==6.0.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==41.0.7
Flask==3.1.1
flask-cors==6.0.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
MarkupSafe==3.0.2
motor==3.7.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
PyJWT==2.10.1
pymongo==4.13.2
PyMySQL==1.1.1
python-dotenv==1.1.1
redis==4.5.4
requests==2.32.4
sentry-sdk[flask]==2.33.2
SQLAlchemy==2.0.41
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
